+++ src_docs/app.component.jsx
@@ -1,11 +1,9 @@
 import React from 'react';
-import { hot } from 'react-hot-loader';
+import { hot } from 'react-hot-loader/root';
 import Layout from './containers/layouts';
 import './app.component.scss';
 import './images/favicon.ico';
 
-export default
-@hot(module)
 class App extends React.Component {
   render() {
     return (
@@ -13,3 +11,5 @@ class App extends React.Component {
     );
   }
 }
+
+export default hot(App);